+++ NEWS.md
@@ -1,5 +1,6 @@
 # backports 1.1.7
 * Added backport for `deparse1()` for R versions prior to 4.0.0.
+* Added backport for `list2DF()` for R versions prior to 4.0.0.
 
 # backports 1.1.6
 * Added backport for `R_user_dir()` for R versions prior to 4.0.0.